+++ SensingKitLib/src/main/java/org/sensingkit/sensingkitlib/SKSensorModuleManager.java
@@ -33,7 +33,7 @@ public class SKSensorModuleManager {
     @SuppressWarnings("unused")
     private static final String TAG = "SKSensorModuleManager";
 
-    private static final int TOTAL_SENSOR_MODULES = 18;
+    private static final int TOTAL_SENSOR_MODULES = 19;
 
     private static SKSensorModuleManager sSensorModuleManager;
     private final Context mApplicationContext;
@@ -196,6 +196,10 @@ public class SKSensorModuleManager {
 
             case HUMIDITY:
                 sensorModule = new SKHumidity(mApplicationContext);
+                break;
+
+            case AIR_PRESSURE:
+                sensorModule = new SKAirPressure(mApplicationContext);
                 break;
 
             // Don't forget the break; here