+++ CHANGELOG.md
@@ -1,7 +1,8 @@
 # Changelog
 
 ### Next
-  - Added support for Humidity sensor module.
+  - Added support for Air Pressure sensor module
+  - Added support for Humidity sensor module
   - Updated Google Play Services to 7.8.0
   - Added support for Android Studio 1.3.1
   - Updated gradle to 1.3.0